Added tests for file I/O errors during a running transfer.
Prerequisite 1) Ability to cause the invocation of data-transfer system calls
such as and sendfile(2) and splice(2) (Linux) to set errno to specific values
and return -1 to indicate an error from client code.

Prerequisite 2) Server needed to be runnable in a thread instead of a process,
otherwise it would not have been possible for test code (in the client process)
to manipulate return and errno values of the server process' invocations of the
data-transfer functions.