Added the ability to cancel transfers of all types.
The reason this needs its own operation (as opposed to, say, simply closing the
destination or status file descriptors) is because of the Send Open File
transaction, in which the server opens the file, sends the file metadata, and
then waits for the client to signal that the file data should be
transferred. I.e., at this point there is no way for the client to signal the
server other than via the request channel: closing the status or destination
file descriptors won't do because the former is the write end of the pipe on the
server (therefore closing it won't result in an event notification on the
server), and it might be undesirable to close the latter as it usually connects
the sendfiled client to one of *its* peers.